Revert "Commit what I have before giving up on insertvalue."
This reverts commit b2af1a3f14171f3f315396ca72fc48307d96329c.

+++ include/pass/LowerExtractvalue.h
@@ -1,11 +1,9 @@
 #pragma once
 
-#include <cstddef>
-
 namespace LL2W {
 	class Function;
 }
 
 namespace LL2W::Passes {
-	size_t lowerExtractvalue(Function &);
+	int lowerExtractvalue(Function &);
 }
